refactor: use djin instead of aoedat

+++ migrations/2021-06-04-105847_create_model/up.sql
@@ -21,83 +21,83 @@ CREATE TABLE help_text
 
 CREATE TABLE civilization
 (
-    id   INT PRIMARY KEY,
+    id   SMALLINT PRIMARY KEY,
     name INT NOT NULL references help_text (id)
 );
 
 CREATE TABLE technology
 (
-    id            INT PRIMARY KEY,
+    id            SMALLINT PRIMARY KEY,
     age           SMALLINT NOT NULL,
     internal_name VARCHAR  NOT NULL,
     name          INT references help_text (id),
-    research_time INT      NOT NULL,
-    wood_cost     INT      NOT NULL,
-    food_cost     INT      NOT NULL,
-    gold_cost     INT      NOT NULL,
-    stone_cost    INT      NOT NULL,
-    is_root BOOLEAN NOT NULL DEFAULT false
+    research_time SMALLINT NOT NULL,
+    wood_cost     SMALLINT NOT NULL,
+    food_cost     SMALLINT NOT NULL,
+    gold_cost     SMALLINT NOT NULL,
+    stone_cost    SMALLINT NOT NULL,
+    is_root       BOOLEAN  NOT NULL DEFAULT false
 );
 
 CREATE TABLE unit
 (
-    id                INT PRIMARY KEY,
+    id                SMALLINT PRIMARY KEY,
     age               SMALLINT NOT NULL,
     unit_type         INT      NOT NULL,
     internal_name     VARCHAR  NOT NULL,
     name              INT references help_text (id),
     help_text_short   INT references help_text (id),
     help_text         INT references help_text (id),
-    wood_cost         INT      NOT NULL,
-    food_cost         INT      NOT NULL,
-    gold_cost         INT      NOT NULL,
-    stone_cost        INT      NOT NULL,
-    attack            INT      NOT NULL,
-    melee_armor       INT      NOT NULL,
-    pierce_armor      INT      NOT NULL,
-    hit_points        INT      NOT NULL,
-    line_of_sight     INT      NOT NULL,
-    garrison_capacity INT      NOT NULL,
-    is_root BOOLEAN NOT NULL DEFAULT false,
-    belongs_to_civ INT REFERENCES civilization(id)
+    wood_cost         SMALLINT NOT NULL,
+    food_cost         SMALLINT NOT NULL,
+    gold_cost         SMALLINT NOT NULL,
+    stone_cost        SMALLINT NOT NULL,
+    attack            SMALLINT NOT NULL,
+    melee_armor       SMALLINT NOT NULL,
+    pierce_armor      SMALLINT NOT NULL,
+    hit_points        SMALLINT NOT NULL,
+    line_of_sight     SMALLINT NOT NULL,
+    garrison_capacity SMALLINT NOT NULL,
+    is_root           BOOLEAN  NOT NULL DEFAULT false,
+    belongs_to_civ    SMALLINT REFERENCES civilization (id)
 );
 
 CREATE TABLE tech_required_tech
 (
-    tech          INT PRIMARY KEY REFERENCES technology (id),
-    required_tech INT NOT NULL REFERENCES technology (id)
+    tech          SMALLINT PRIMARY KEY REFERENCES technology (id),
+    required_tech SMALLINT NOT NULL REFERENCES technology (id)
 );
 
 CREATE TABLE unit_required_tech
 (
-    unit          INT PRIMARY KEY REFERENCES unit (id),
-    required_tech INT NOT NULL REFERENCES technology (id)
+    unit          SMALLINT PRIMARY KEY REFERENCES unit (id),
+    required_tech SMALLINT NOT NULL REFERENCES technology (id)
 );
 
 CREATE TABLE unit_required_unit
 (
-    unit          INT PRIMARY KEY REFERENCES unit (id),
-    required_unit INT NOT NULL REFERENCES unit (id)
+    unit          SMALLINT PRIMARY KEY REFERENCES unit (id),
+    required_unit SMALLINT NOT NULL REFERENCES unit (id)
 );
 
 CREATE TABLE tech_required_unit
 (
-    tech INT PRIMARY KEY REFERENCES technology (id),
-    required_unit INT NOT NULL REFERENCES unit (id)
+    tech          SMALLINT PRIMARY KEY REFERENCES technology (id),
+    required_unit SMALLINT NOT NULL REFERENCES unit (id)
 );
 
 CREATE TABLE civ_tech
 (
-    civ_id INT NOT NULL REFERENCES civilization (id),
-    tech_id INT NOT NULL REFERENCES technology (id),
-    PRIMARY KEY(civ_id, tech_id)
+    civ_id  SMALLINT NOT NULL REFERENCES civilization (id),
+    tech_id SMALLINT NOT NULL REFERENCES technology (id),
+    PRIMARY KEY (civ_id, tech_id)
 );
 
 CREATE TABLE civ_unit
 (
-    civ_id INT NOT NULL REFERENCES civilization (id),
-    unit_id INT NOT NULL REFERENCES unit (id),
-    PRIMARY KEY(civ_id, unit_id)
+    civ_id  SMALLINT NOT NULL REFERENCES civilization (id),
+    unit_id SMALLINT NOT NULL REFERENCES unit (id),
+    PRIMARY KEY (civ_id, unit_id)
 );
